get everything working using a customary layout,
so i can use this on my western keyboard. basically, i use space
for the same-shift thumb key (since it's easy to hit with the same
hand), whereas i use the buttons left and right of space for
cross-shifting, since the other hand (that is not hitting a button)
is a bit more flexible with its thumb.